Refactored all the files, added code to strip prefix r/ from the added data, and added r/ prefix for display in adapters. added prefix stripping function for instant add function in searchActivity
