+++ app/build.gradle
@@ -20,6 +20,7 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
     implementation 'com.android.support:appcompat-v7:27.0.2'
     implementation 'com.squareup.okhttp3:okhttp:3.9.1'
     implementation 'com.android.support:recyclerview-v7:27.0.2'
