+++ app/src/main/java/com/example/redditimages/redditgram/Adapters/FeedListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.redditimages.redditgram;
+package com.example.redditimages.redditgram.Adapters;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -7,6 +7,7 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.example.redditimages.redditgram.R;
 import com.example.redditimages.redditgram.Utils.DownloadImageTask;
 import com.example.redditimages.redditgram.Utils.FeedFetchUtils;
