+++ app/src/main/java/com/example/redditimages/redditgram/Fragments/SettingsFragment.java
@@ -1,9 +1,11 @@
-package com.example.redditimages.redditgram;
+package com.example.redditimages.redditgram.Fragments;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.support.v7.preference.EditTextPreference;
 import android.support.v7.preference.PreferenceFragmentCompat;
+
+import com.example.redditimages.redditgram.R;
 
 /**
  * Created by poj on 3/15/18.
