+++ app/src/main/java/com/example/redditimages/redditgram/SubredditActivity.java
@@ -1,6 +1,5 @@
 package com.example.redditimages.redditgram;
 
-import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.support.v7.app.AppCompatActivity;
@@ -9,20 +8,22 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 
+import com.example.redditimages.redditgram.Adapters.SubredditAdapter;
 import com.example.redditimages.redditgram.SubredditDB.SubredditContract;
 import com.example.redditimages.redditgram.SubredditDB.SubredditDBHelper;
-import com.example.redditimages.redditgram.Utils.SubredditSearchUtils;
 
 import java.util.ArrayList;
 
 public class SubredditActivity extends AppCompatActivity implements SubredditAdapter.OnSubredditItemClickListener{
+
+    private static final String TAG = SubredditActivity.class.getSimpleName();
 
     private SubredditAdapter mAdapter;
     private RecyclerView mSubredditListItemsRV;
     private SQLiteDatabase mDB;
     public ArrayList<String> subredditItems;
 
-    private static final String TAG = SubredditActivity.class.getSimpleName();
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -47,7 +48,7 @@ public class SubredditActivity extends AppCompatActivity implements SubredditAda
 
 
 
-    //onclicklistener
+    /* onclicklistener */
 
     @Override
     public void onSubredditItemClick(String subredditItem) {
@@ -56,59 +57,36 @@ public class SubredditActivity extends AppCompatActivity implements SubredditAda
 
     public long deleteSubredditFromDB(String subredditName) {
         if (subredditName != null) {
-            String sqlSelection = SubredditContract.FollowingSubreddits.COLUMN_SUBREDDIT_NAME + " = ?";
+            String sqlSelection = SubredditContract.SavedSubreddits.COLUMN_SUBREDDIT_NAME + " = ?";
             String[] sqlSelectionArgs = {subredditName};
-            return mDB.delete(SubredditContract.FollowingSubreddits.TABLE_NAME, sqlSelection, sqlSelectionArgs);
+            return mDB.delete(SubredditContract.SavedSubreddits.TABLE_NAME, sqlSelection, sqlSelectionArgs);
         } else {
             Log.d(TAG, "Failed to remove subreddit from database.");
             return -1;
         }
     }
 
-    private boolean checkIAdded(String subredditName) {
-        boolean isMarked = false;
-        if (subredditName != null) {
-            String sqlSelection =
-                    SubredditContract.FollowingSubreddits.COLUMN_SUBREDDIT_NAME  + " = ?";
-            String[] sqlSelectionArgs = { subredditName };
-            Cursor cursor = mDB.query(
-                    SubredditContract.FollowingSubreddits.TABLE_NAME,
-                    null,
-                    sqlSelection,
-                    sqlSelectionArgs,
-                    null,
-                    null,
-                    null
-            );
-            isMarked = cursor.getCount() > 0;
-            cursor.close();
-        }
-        return isMarked;
-    }
 
     private ArrayList<String> getAllSubredditsFromDB() {
         Cursor cursor = mDB.query(
-                 SubredditContract.FollowingSubreddits.TABLE_NAME,
+                 SubredditContract.SavedSubreddits.TABLE_NAME,
                 null,
                 null,
                 null,
                 null,
                 null,
-                SubredditContract.FollowingSubreddits.COLUMN_SUBREDDIT_NAME + " ASC"
+                SubredditContract.SavedSubreddits.COLUMN_SUBREDDIT_NAME + " ASC"
         );
 
         ArrayList<String> subredditResults = new ArrayList<>();
         while (cursor.moveToNext()) {
             String searchResult;
             searchResult= cursor.getString(
-                    cursor.getColumnIndex(SubredditContract.FollowingSubreddits.COLUMN_SUBREDDIT_NAME)
+                    cursor.getColumnIndex(SubredditContract.SavedSubreddits.COLUMN_SUBREDDIT_NAME)
             );
             subredditResults.add(searchResult);
         }
         cursor.close();
         return subredditResults;
     }
-
-
-
 }
