+++ app/src/main/res/layout/activity_settings.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:name="com.example.redditimages.redditgram.SettingsFragment"
+    android:name="com.example.redditimages.redditgram.Fragments.SettingsFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".SettingsActivity"/>
