+++ app/src/main/res/layout/search_subreddit_item.xml
@@ -1,35 +1,49 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
 
-    <TextView
-        android:id="@+id/tv_search_subreddit_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="22sp"
-        android:padding="16dp"
-        android:layout_alignParentLeft="true"/>
-
-    <LinearLayout
-        android:orientation="horizontal"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentRight="true">
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
 
         <TextView
-            android:id="@+id/tv_search_subreddit_category"
+            android:id="@+id/tv_search_subreddit_name"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textSize="22sp"
-            android:padding="16dp" />
+            android:padding="16dp"
+            android:layout_alignParentLeft="true"/>
 
-        <ImageButton
-            android:id="@+id/ib_add_subreddit_button"
-            android:src="@drawable/ic_action_add"
+        <LinearLayout
+            android:orientation="horizontal"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
+            android:layout_height="wrap_content"
+            android:layout_alignParentRight="true">
 
-    </LinearLayout>
+            <TextView
+                android:id="@+id/tv_search_subreddit_category"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textSize="22sp"
+                android:padding="16dp" />
 
-</RelativeLayout>
+            <ImageButton
+                android:id="@+id/ib_add_subreddit_button"
+                android:src="@drawable/ic_action_add"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content" />
+
+        </LinearLayout>
+
+    </RelativeLayout>
+
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginLeft="8dp"
+        android:layout_marginRight="8dp"
+        android:background="#dadada" />
+
+</LinearLayout>
